backend-rag-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
qdrant-client==1.7.0
huggingface-hub==0.16.4
sentence-transformers==2.2.2
numpy==1.26.3
prometheus-client==0.19.0
loguru==0.7.2
pytest==7.4.0
redis==5.0.1
openai==1.12.0
